Pixel Aspect Ratio Patch by carl edman
If transcoding to an S3 is required, PAR is changed to 1:1 if
ffmpeg reports the PAR correctly and PAR is not already 1:1.

This is essentially the foundation of the PAR patch
submitted by Carl Edman.  I have removed the config setting
and tivo_compatible par 1:1 check because I want to change the
functionality to work a bit differently than submitted.